+++ RememberMePlusApp/Infrastructure/Data/UnitOfWorkFactory.cs
@@ -5,8 +5,8 @@ public sealed class UnitOfWorkFactory(IDbConnectionFactory dbConnectionFactory) 
     private readonly IDbConnectionFactory _dbConnectionFactory = dbConnectionFactory;
 
     /// <summary>
-    /// Crea una nueva unidad de trabajo con conexión y transacción abiertas.
+    /// Crea una nueva unidad de trabajo con conexión abierta y transacción opcional.
     /// </summary>
-    public async Task<IUnitOfWork> CreateAsync(CancellationToken cancellationToken = default)
-        => await UnitOfWork.CreateAsync(_dbConnectionFactory, cancellationToken);
+    public async Task<IUnitOfWork> CreateAsync(bool useTransaction = true, CancellationToken cancellationToken = default)
+        => await UnitOfWork.CreateAsync(_dbConnectionFactory, useTransaction, cancellationToken);
 }